Refactor _pick_id_dtype_for_newick to use np.min_scalar_type
Replace manual iinfo threshold checks with np.min_scalar_type, then
promote the unsigned result to its signed counterpart (uint8->int16,
uint16->int32, uint32->int64, uint64->int64).

+++ tests/test_phyloframe/test_legacy/test_alifestd_from_newick_polars.py
@@ -259,21 +259,21 @@ def test_id_dtype_int32():
 
 def test_id_dtype_none_small():
     result = alifestd_from_newick_polars("(A,B);", id_dtype=None)
-    # 1 comma -> fits in int8
-    assert result["id"].dtype == pl.Int8
-    assert result["ancestor_id"].dtype == pl.Int8
+    # 1 comma -> uint8 via min_scalar_type -> int16 signed
+    assert result["id"].dtype == pl.Int16
+    assert result["ancestor_id"].dtype == pl.Int16
     assert len(result) == 3
 
 
 def test_id_dtype_none_empty():
     result = alifestd_from_newick_polars("", id_dtype=None)
-    assert result["id"].dtype == pl.Int8
+    assert result["id"].dtype == pl.Int16
     assert len(result) == 0
 
 
 def test_id_dtype_none_values_correct():
     result = alifestd_from_newick_polars("(A,(B,C));", id_dtype=None)
-    assert result["id"].dtype == pl.Int8
+    assert result["id"].dtype == pl.Int16
     assert len(result) == 5
     root = result.filter(pl.col("ancestor_id") == pl.col("id"))
     assert len(root) == 1
